Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,13 +1,13 @@
 #include <stdio.h>
 
 int main () {
-    int num1 ;
-    int num2 ;
-    int num3 ;
+    float num1 ;
+    float num2 ;
+    float num3 ;
     scanf("%d", &num1);
     scanf("%d", &num2);
     scanf("%d", &num3);
-    int sum = num1 + num2 + num3 ;
+    float sum = num1 + num2 + num3 ;
     float average = sum/3;
     printf("Average: %.2f\n", average);
     return 0;
